New figures, max_features of rf tuned, highly corr variables removed
Was able to tune and get a decrease in RMSE by implementing changes in max_features paramter in building the RF models. Something to consider is taking outliers out of the model. Also highly correlated features were removed to get a more accurate picture of variable importance, since for the N variables they weren't seen as individually important for the model since they were all correlated. But when just NH4 is left and other N removed, the model performs slightly better and gives more importance to NH4.